Slight modifications to prev commit.
Instead of having various Words objects around the
classes we want to have just one in App. This allows
all the locations to be stored in the same map and
also means we can access this map. Not such a problem
for menus as every menus instance has a map with every
item and its price. Thinking about implementing a similar
system for Words but not sure if reading all the words from
the directory using the solution I found on SO is too much
like plagarism.