Updated GeoJsonParsing and Words.
Added a method 'parsePointsToW3w' to GeoJsonParsing which
takes a GeoJsonString and from it can extrapolate a property
called 'coordinates'. This allows us to get the w3w address for
each land mark so they can be added to wordsMap.

Also 'checkIsWhatThreeWordsAddress' to Words. This simply allows us
to check if a given string is of what3words format by using regular
expressions.