Can now find shortest path for an order in terms of w3w addresses.
Also began working on implementing the Drone class.
Now need to get the drone to actually move along each of the sub paths
from its start location, to shop(s), to delivery by using its real time
location. Once that is complete, we need to add functionality to store the
moves in the database. After that, we're pretty much done.