Fixed small bug in Words.buildGraphFromWords
We were getting every pair of entries possible in wordsMap, even
when the 2 entries were the same. This meant our graph was containing
a loop for each location which isn't allowed so we added a small
check in line 113 in Words to fix this.

Now fairly confident that our graphs are being built successfully.